% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect_idempotent.R
\name{expect_idempotent}
\alias{expect_idempotent}
\title{Do you expect a function to be idempotent?}
\usage{
expect_idempotent(f, x, ...)
}
\arguments{
\item{f}{function. The function you are testing to be idempotent}

\item{x}{parameter(s). The value under which you test the function.}

\item{...}{Arguments passed into \code{\link[waldo:compare]{waldo::compare()}}}
}
\description{
This function tests whether if you apply a function on a value
twice, it is the same as applying it a single time.
}
\details{
Tests that \eqn{f(f(x)) = f(x)}
}
\examples{
x <- c(2, 1, 3)
expect_idempotent(mean, x)
expect_idempotent(sort, x)
}
